Added piece to change java to version 1.8 for Gremlin to run
Because of two versions of JDK on my computer, it defaults to the more recent version. To change my current shell's environment, I'll need to run this code every time. I could uninstall the more recent version, or change some settings more in-depth to make v. 1.8 the default, but doing it this way for now.